+++ src/main/java/com/example/demo/Repository/CupomRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface CupomRepository extends JpaRepository<Cupom, BigInteger> {
+public interface CupomRepository extends JpaRepository<Cupom, Long> {
 }
