+++ src/main/java/com/example/demo/Repository/ProdutoRepository.java
@@ -4,5 +4,5 @@ import com.example.demo.Model.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 
-public interface ProdutoRepository extends JpaRepository<Produto, BigInteger> {
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 }
