+++ src/main/java/com/example/demo/Service/CupomService.java
@@ -6,9 +6,9 @@ import java.math.BigInteger;
 import java.util.List;
 
 public interface CupomService {
-    public Cupom getById(BigInteger id);
+    public Cupom getById(Long id);
     public List<Cupom> getAll();
     public Cupom save(Cupom cupom);
-    public Cupom update(BigInteger id, Cupom cupom);
-    public void delete(BigInteger id);
+    public Cupom update(Long id, Cupom cupom);
+    public void delete(Long id);
 }
