+++ src/main/java/com/example/demo/Service/ProdutoService.java
@@ -6,9 +6,9 @@ import java.math.BigInteger;
 import java.util.List;
 
 public interface ProdutoService {
-    public Produto getById(BigInteger id);
+    public Produto getById(Long id);
     public List<Produto> getAll();
     public Produto save(Produto produto);
-    public Produto update(BigInteger id, Produto produto);
-    public void delete(BigInteger id);
+    public Produto update(Long id, Produto produto);
+    public void delete(Long id);
 }
