+++ src/main/java/com/example/demo/Service/UsuarioService.java
@@ -6,9 +6,9 @@ import java.math.BigInteger;
 import java.util.List;
 
 public interface UsuarioService {
-    public Usuario getById(BigInteger id);
+    public Usuario getById(Long id);
     public List<Usuario> getAll();
     public Usuario save(Usuario usuario);
-    public Usuario update(BigInteger id, Usuario usuario);
-    public void delete(BigInteger id);
+    public Usuario update(Long id, Usuario usuario);
+    public void delete(Long id);
 }
